lib: enable non ARM build for libs
Enable non ARM builds for libs and apps. When the platform is not ARM,
the build is treated as host build and only the libs & apps required to
be run on host would be build. If the platform is ARM but the build is
for host, then the new option ('enable_host_build') can be used.

Signed-off-by: Anoob Joseph <anoobj@marvell.com>
Change-Id: I49c9f46e4aac1c4f5e418dc559e46fab6550b43b
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpu-offload/+/142460
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
(cherry picked from commit 17176230fabdce37503f6c62e35e29c50707e182)
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpu-offload/+/143762
Reviewed-by: Jerin Jacob <jerinj@marvell.com>